+++ skoolkit/config.py
@@ -46,6 +46,7 @@ COMMANDS = {
     'rzxplay': {
         'Screen': (1, 'screen'),
         'ScreenFps': (50, 'fps'),
+        'ScreenScale': (2, 'scale'),
         'TraceHeader': ('', ''),
         'TraceLine': ('F:{fr:0{fw}} C:{fc:05} I:{rr:05} ${pc:04X} {i}', ''),
         'TraceOperand': ('$,02X,04X', ''),